+++ src/app/layout.tsx
@@ -1,7 +1,12 @@
 import type { Metadata } from "next";
 import "./globals.css";
+import { Baskervville } from "next/font/google";
 import Footer from "@/components/layout/Footer";
 import Topbar from "@/components/layout/Topbar";
+
+const baskervville = Baskervville({
+  subsets: ["latin"],
+});
 
 export const metadata: Metadata = {
   title: "Haidar's Folio",
@@ -14,7 +19,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en">
+    <html lang="en" className={baskervville.className}>
       <body>
         <Topbar />
         {children}