+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.5] - 2025-11-08
+
+### Added
+
+- Welcome and intro in landing section.
+- `h-header-height` value in `globals.css`.
+
+### Changed
+
+- Now using Baskervville font by ANRT as default font (font set in `layout.tsx`).
+- Replaced `<div>` wrapper for showcase video with a shadcn/ui Aspect Ratio component.
+
+## [0.0.4] - 2025-11-04
 
 ### Added